Test was not able to run on not installed instance
