+++ test/unit/model/taskQueue/QueueTest.php
@@ -108,83 +108,6 @@ class QueueTest extends TestCase
         ];
     }
 
-    /**
-     * @dataProvider provideDequeueOptions
-     */
-    public function testDequeueWhenTaskPoppedOrNot($dequeuedElem, $expected)
-    {
-        $lockMock = $this->getMockBuilder(LockInterface::class)->disableOriginalConstructor()->getMock();
-        $lockMock->method('acquire')->willReturn(true);
-        $lockMock->method('release')->willReturn(true);
-        /** @var QueueBrokerInterface|MockObject $queueBrokerMock */
-        $queueBrokerMock = $this->getMockBuilder(QueueBrokerInterface::class)
-            ->disableOriginalConstructor()
-            ->setMethods(['pop', 'setServiceLocator'])
-            ->getMockForAbstractClass();
-
-        $queueBrokerMock
-            ->method('pop')
-            ->willReturn($dequeuedElem);
-
-        $queueName = 'name of the queue';
-        $subject = $this->getMockBuilder(Queue::class)
-            ->setConstructorArgs([$queueName, $queueBrokerMock])
-            ->setMethods(['createLock'])
-            ->getMock();
-        $subject->method('createLock')
-            ->willReturn($lockMock);
-
-        if ($dequeuedElem instanceof AbstractTask) {
-            /** @var TaskLogInterface|MockObject $taskLogMock */
-            $taskLogMock = $this->getMockForAbstractClass(TaskLogInterface::class);
-            $taskLogMock
-                ->method('getStatus')
-                ->willReturnArgument(0);
-
-            if ($dequeuedElem->getId() !== TaskLogInterface::STATUS_CANCELLED) {
-                $taskLogMock->expects($this->once())
-                    ->method('setStatus')
-                    ->with($dequeuedElem->getId(), TaskLogInterface::STATUS_DEQUEUED);
-
-                /** @var LoggerInterface|MockObject $loggerMock */
-                $loggerMock = $this->getMockBuilder(LoggerInterface::class)
-                    ->disableOriginalConstructor()
-                    ->setMethods(['info'])
-                    ->getMockForAbstractClass();
-                $loggerMock
-                    ->expects($this->once())
-                    ->method('info')
-                    ->with(
-                        sprintf('Task %s has been dequeued', $dequeuedElem->getId()),
-                        [
-                            'PID' => getmypid(),
-                            'QueueName' => $queueName,
-                        ]
-                    );
-
-                $subject->setLogger($loggerMock);
-            }
-
-            $subject->setTaskLog($taskLogMock);
-        }
-
-        $this->assertEquals($expected, $subject->dequeue());
-    }
-
-    public function provideDequeueOptions()
-    {
-        $validId = 'a valid id';
-
-        $canceledTask = $this->createTaskMock(TaskLogInterface::STATUS_CANCELLED);
-        $dequeuedTask = $this->createTaskMock($validId);
-
-        return [
-            'empty queue' => [null, null],
-            'canceled task' => [$canceledTask, $canceledTask],
-            'dequeued task' => [$dequeuedTask, $dequeuedTask],
-        ];
-    }
-
     public function createTaskMock($id)
     {
         $taskMock = $this->getMockBuilder(AbstractTask::class)
